requirements.txt
backcall==0.1.0
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
decorator==4.3.0
Django==2.0.3
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-webpack-loader==0.6.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
Flask==0.12.2
idna==2.6
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
itsdangerous==0.24
jdcal==1.4
jedi==0.12.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.12
numpy==1.14.3
openpyxl==2.5.3
parso==0.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psutil==5.4.6
py==1.5.2
Pygments==2.2.0
PyJWT==1.6.4
pytest==3.2.5
pytz==2018.3
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
terminaltables==3.1.0
traitlets==4.3.2
urllib3==1.22
virtualenv==16.0.0
virtualenvwrapper-win==1.2.5
wcwidth==0.1.7
Werkzeug==0.14.1
